+++ LookupAnything/Framework/Targets/ObjectTarget.cs
@@ -1,7 +1,8 @@
-﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
 using StardewValley;
+using StardewValley.Objects;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Targets
 {
@@ -33,6 +34,8 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Targets
         {
             Object obj = (Object)this.Value;
             Rectangle boundingBox = obj.getBoundingBox(this.GetTile());
+            if (obj is Furniture furniture)
+                return this.GetSpriteArea(boundingBox, furniture.sourceRect);
             if (obj.bigCraftable)
                 return this.GetSpriteArea(boundingBox, Object.getSourceRectForBigCraftable(obj.parentSheetIndex));
             if (obj is Fence fence)
@@ -52,15 +55,20 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Targets
             // get sprite data
             Texture2D spriteSheet;
             Rectangle sourceRectangle;
-            if (obj.bigCraftable)
+            if (obj is Furniture furniture)
             {
-                spriteSheet = Game1.bigCraftableSpriteSheet;
-                sourceRectangle = Object.getSourceRectForBigCraftable(obj.parentSheetIndex);
+                spriteSheet = Furniture.furnitureTexture;
+                sourceRectangle = furniture.sourceRect;
             }
             else if (obj is Fence fence)
             {
-                spriteSheet = this.Reflection.GetPrivateValue<Texture2D>(obj, "fenceTexture");
+                spriteSheet = this.Reflection.GetField<Texture2D>(obj, "fenceTexture").GetValue();
                 sourceRectangle = this.GetSourceRectangle(fence, Game1.currentLocation);
+            }
+            else if (obj.bigCraftable)
+            {
+                spriteSheet = Game1.bigCraftableSpriteSheet;
+                sourceRectangle = Object.getSourceRectForBigCraftable(obj.parentSheetIndex);
             }
             else
             {
@@ -117,7 +125,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Targets
                     spriteID = Fence.fenceDrawGuide[index];
             }
 
-            Texture2D texture = this.Reflection.GetPrivateValue<Texture2D>(fence, "fenceTexture");
+            Texture2D texture = this.Reflection.GetField<Texture2D>(fence, "fenceTexture").GetValue();
             return new Rectangle(spriteID * Fence.fencePieceWidth % texture.Bounds.Width, spriteID * Fence.fencePieceWidth / texture.Bounds.Width * Fence.fencePieceHeight, Fence.fencePieceWidth, Fence.fencePieceHeight);
         }
     }