+++ TractorMod/ModEntry.cs
@@ -83,9 +83,10 @@ namespace Pathoschild.Stardew.TractorMod
             this.Config = helper.ReadConfig<ModConfig>();
             this.Attachments = new IAttachment[]
             {
-                new CustomToolAttachment(this.Config), // should be first so it can override tools
+                new CustomAttachment(this.Config), // should be first so it can override default attachments
                 new AxeAttachment(this.Config),
                 new FertilizerAttachment(),
+                new GrassStarterAttachment(),
                 new HoeAttachment(),
                 new PickaxeAttachment(this.Config),
                 new ScytheAttachment(),
@@ -157,7 +158,8 @@ namespace Pathoschild.Stardew.TractorMod
                 if (e.NewMenu is CarpenterMenu)
                 {
                     this.Helper.Reflection
-                        .GetPrivateValue<List<BluePrint>>(e.NewMenu, "blueprints")
+                        .GetField<List<BluePrint>>(e.NewMenu, "blueprints")
+                        .GetValue()
                         .Add(this.GetBlueprint());
                 }
 
@@ -165,13 +167,14 @@ namespace Pathoschild.Stardew.TractorMod
                 else if (this.IsPelicanFiberLoaded && e.NewMenu.GetType().FullName == this.PelicanFiberMenuFullName)
                 {
                     this.Helper.Reflection
-                        .GetPrivateValue<List<BluePrint>>(e.NewMenu, "Blueprints")
+                        .GetField<List<BluePrint>>(e.NewMenu, "Blueprints")
+                        .GetValue()
                         .Add(this.GetBlueprint());
                 }
                 else if (this.IsFarmExpansionLoaded && e.NewMenu.GetType().FullName == this.FarmExpansionMenuFullName)
                 {
                     this.Helper.Reflection
-                        .GetPrivateMethod(e.NewMenu, "AddFarmBluePrint")
+                        .GetMethod(e.NewMenu, "AddFarmBluePrint")
                         .Invoke(this.GetBlueprint());
                 }
             }
@@ -183,7 +186,7 @@ namespace Pathoschild.Stardew.TractorMod
         private void InputEvents_ButtonPressed(object sender, EventArgsInput e)
         {
             // summon tractor
-            if (this.Config.Controls.SummonTractor.Contains(e.Button))
+            if (Context.IsPlayerFree && this.Config.Controls.SummonTractor.Contains(e.Button))
                 this.Tractor?.SetLocation(Game1.currentLocation, Game1.player.getTileLocation());
         }
 
@@ -326,8 +329,8 @@ namespace Pathoschild.Stardew.TractorMod
                     {
                         new FarmerSprite.AnimationFrame(24, 75),
                         new FarmerSprite.AnimationFrame(25, 75),
-                        new FarmerSprite.AnimationFrame(26, 300, false, false, farmer => this.Helper.Reflection.GetPrivateMethod(robin,"robinHammerSound").Invoke(farmer)),
-                        new FarmerSprite.AnimationFrame(27, 1000, false, false, farmer => this.Helper.Reflection.GetPrivateMethod(robin,"robinVariablePause").Invoke(farmer))
+                        new FarmerSprite.AnimationFrame(26, 300, false, false, farmer => this.Helper.Reflection.GetMethod(robin,"robinHammerSound").Invoke(farmer)),
+                        new FarmerSprite.AnimationFrame(27, 1000, false, false, farmer => this.Helper.Reflection.GetMethod(robin,"robinVariablePause").Invoke(farmer))
                     });
                     robin.ignoreScheduleToday = true;
                     Game1.warpCharacter(robin, location.Name, new Vector2(garage.tileX + garage.tilesWide / 2, garage.tileY + garage.tilesHigh / 2), false, false);