+++ ChestsAnywhere/release-notes.md
@@ -6,6 +6,7 @@
 * Fixed shipping bin UI not allowing more than 36 items.
 * Fixed shipping bin UI resetting gamepad cursor position on refresh.
 * Fixed error when the range is set to `None` and a chest is opened directly.
+* Improved German translations. (Thanks to Husky110!)
 
 ## 1.12.2
 * Fixed broken shipping bin UI.