+++ DataMaps/release-notes.md
@@ -1,4 +1,8 @@
 # Release notes
+## 1.2.1
+* Fixed error in the Cobalt integration.
+* Added Chinese and German translations. (Thanks to yuwenlan and Husky110!)
+
 ### 1.2
 * Added: point at a scarecrow/sprinkler/etc in its data map to highlight that one's range.
 * Added: two overlapping groups of the same color will now share one border (configurable).